+++ bigmler/tests/test_19_mlm_prediction.py
@@ -46,13 +46,14 @@ class TestMLM(object):
         """Calling generic teardown for every method
 
         """
+        print "\nEnd of tests in: %s\n-------------------\n" % __name__
         teardown_class()
 
     def setup(self):
-        """No setup operations for every method at present
-
         """
-        pass
+            Debug information
+        """
+        print "\n-------------------\nTests in: %s\n" % __name__
 
     def setup_scenario1(self):
         """